+++ src/lib.rs
@@ -27,6 +27,7 @@
     // clippy::indexing_slicing,
     // clippy::multiple_crate_versions,
     // clippy::multiple_inherent_impl,
+    // clippy::missing_const_for_fn,
     clippy::cast_lossless,
     clippy::cast_possible_truncation,
     clippy::cast_possible_wrap,
@@ -53,7 +54,6 @@
     clippy::match_same_arms,
     clippy::maybe_infinite_iter,
     clippy::mem_forget,
-    clippy::missing_const_for_fn,
     clippy::module_name_repetitions,
     clippy::mut_mut,
     clippy::needless_borrow,
@@ -88,6 +88,20 @@ mod inbound;
 mod outbound;
 mod parser;
 mod shared_state;
+
+#[cfg(feature = "fault_injection")]
+mod fault_injection;
+
+#[cfg(feature = "fault_injection")]
+use fault_injection::{inject_delay, inject_io_failure};
+
+#[cfg(not(feature = "fault_injection"))]
+fn inject_delay() {}
+
+#[cfg(not(feature = "fault_injection"))]
+fn inject_io_failure() -> io::Result<()> {
+    Ok(())
+}
 
 /// Functionality relating to TLS configuration
 pub mod tls;
@@ -167,21 +181,8 @@ impl ServerInfo {
     }
 }
 
-/// A `ConnectionStatus` describes the current sub-status of a `Connection`.
-#[derive(Debug, PartialEq, Clone, Copy)]
-enum ConnectionStatus {
-    /// An established connection.
-    Connected,
-    /// A permanently closed connection.
-    Closed,
-    /// A connection that has lost connectivity, but may reestablish connectivity.
-    Disconnected,
-    /// A connection in the process of reestablishing connectivity after a disconnect.
-    Reconnecting,
-}
-
 #[derive(Default, Clone)]
-pub(crate) struct Callback(Option<Arc<dyn Fn(&ServerInfo) + Send + Sync + 'static>>);
+pub(crate) struct Callback(Option<Arc<dyn Fn() + Send + Sync + 'static>>);
 
 impl fmt::Debug for Callback {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
@@ -229,6 +230,7 @@ pub struct ConnectionOptions<TypeState> {
     reconnect_buffer_size: usize,
     disconnect_callback: Callback,
     reconnect_callback: Callback,
+    close_callback: Callback,
     tls_connector: Option<tls::TlsConnector>,
     tls_required: bool,
 }
@@ -244,6 +246,7 @@ impl Default for ConnectionOptions<options_typestate::NoAuth> {
             max_reconnects: Some(60),
             disconnect_callback: Callback(None),
             reconnect_callback: Callback(None),
+            close_callback: Callback(None),
             tls_connector: None,
             tls_required: false,
         }
@@ -260,6 +263,7 @@ impl<T> fmt::Debug for ConnectionOptions<T> {
             .entry(&"max_reconnects", &self.max_reconnects)
             .entry(&"disconnect_callback", &self.disconnect_callback)
             .entry(&"reconnect_callback", &self.reconnect_callback)
+            .entry(&"close_callback", &self.close_callback)
             .entry(
                 &"tls_connector",
                 if self.tls_connector.is_some() {
@@ -305,6 +309,7 @@ impl ConnectionOptions<options_typestate::NoAuth> {
             typestate: PhantomData,
             no_echo: self.no_echo,
             name: self.name,
+            close_callback: self.close_callback,
             disconnect_callback: self.disconnect_callback,
             reconnect_callback: self.reconnect_callback,
             reconnect_buffer_size: self.reconnect_buffer_size,
@@ -336,6 +341,7 @@ impl ConnectionOptions<options_typestate::NoAuth> {
             no_echo: self.no_echo,
             name: self.name,
             reconnect_buffer_size: self.reconnect_buffer_size,
+            close_callback: self.close_callback,
             disconnect_callback: self.disconnect_callback,
             reconnect_callback: self.reconnect_callback,
             max_reconnects: self.max_reconnects,
@@ -403,6 +409,8 @@ impl<TypeState> ConnectionOptions<TypeState> {
     /// when accepting outgoing traffic in disconnected
     /// mode.
     ///
+    /// The default value is 8mb.
+    ///
     /// # Example
     /// ```
     /// # fn main() -> std::io::Result<()> {
@@ -439,6 +447,7 @@ impl<TypeState> ConnectionOptions<TypeState> {
             max_reconnects: self.max_reconnects,
             disconnect_callback: self.disconnect_callback,
             reconnect_callback: self.reconnect_callback,
+            close_callback: self.close_callback,
             tls_connector: self.tls_connector,
             tls_required: self.tls_required,
             // move options into the Finalized state by setting
@@ -462,7 +471,7 @@ impl<TypeState> ConnectionOptions<TypeState> {
     /// a server has been lost.
     pub fn set_disconnect_callback<F>(mut self, cb: F) -> Self
     where
-        F: Fn(&ServerInfo) + Send + Sync + 'static,
+        F: Fn() + Send + Sync + 'static,
     {
         self.disconnect_callback = Callback(Some(Arc::new(cb)));
         self
@@ -472,9 +481,20 @@ impl<TypeState> ConnectionOptions<TypeState> {
     /// server has been established.
     pub fn set_reconnect_callback<F>(mut self, cb: F) -> Self
     where
-        F: Fn(&ServerInfo) + Send + Sync + 'static,
+        F: Fn() + Send + Sync + 'static,
     {
         self.disconnect_callback = Callback(Some(Arc::new(cb)));
+        self
+    }
+
+    /// Set a callback to be executed when the client has been
+    /// closed due to exhausting reconnect retries to known servers
+    /// or by completing a drain request.
+    pub fn set_close_callback<F>(mut self, cb: F) -> Self
+    where
+        F: Fn() + Send + Sync + 'static,
+    {
+        self.close_callback = Callback(Some(Arc::new(cb)));
         self
     }
 
@@ -543,16 +563,36 @@ pub(crate) struct ShutdownDropper {
 
 impl Drop for ShutdownDropper {
     fn drop(&mut self) {
-        self.shared_state.shut_down();
+        self.shared_state.shutdown();
+
+        inject_delay();
+        if let Some(mut threads) = self.shared_state.threads.lock().take() {
+            let inbound = threads.inbound.take().unwrap();
+            let outbound = threads.outbound.take().unwrap();
+
+            if let Err(error) = inbound.join() {
+                log::error!("error encountered in inbound thread: {:?}", error);
+            }
+            if let Err(error) = outbound.join() {
+                log::error!("error encountered in outbound thread: {:?}", error);
+            }
+        }
     }
 }
 
 /// A NATS connection.
 #[derive(Debug, Clone)]
 pub struct Connection {
-    shared_state: Arc<SharedState>,
-    shutdown_dropper: Arc<ShutdownDropper>,
     sid: Arc<AtomicUsize>,
+    shared_state: Arc<SharedState>,
+    // we split the `ShutdownDropper` into
+    // a separate Arc from the `SharedState`
+    // because the `ShutdownDropper` will only
+    // be held by "user-facing" structures, and
+    // the `SharedState` may be held by background
+    // threads that we wish to terminate once
+    // all of the user-held structures are destroyed.
+    shutdown_dropper: Arc<ShutdownDropper>,
 }
 
 #[derive(Serialize, Clone, Debug)]
@@ -857,6 +897,7 @@ impl Connection {
         let (s, r) = crossbeam_channel::bounded(1);
 
         // We take out the mutex before sending a ping (which may fail)
+        inject_delay();
         let mut pongs = self.shared_state.pongs.lock();
 
         // This will throw an error if the system is disconnected.
@@ -882,19 +923,25 @@ impl Connection {
         Ok(())
     }
 
-    /// Close a NATS connection.
+    /// Close a NATS connection. All clones of
+    /// this `Connection` will also be closed,
+    /// as the backing IO threads are shared.
+    ///
+    /// If the client is currently connected
+    /// to a server, the outbound write buffer
+    /// will be flushed in the process of
+    /// shutting down.
     ///
     /// # Example
     /// ```
     /// # fn main() -> std::io::Result<()> {
     /// # let nc = nats::connect("demo.nats.io")?;
-    /// nc.close()?;
+    /// nc.close();
     /// # Ok(())
     /// # }
     /// ```
-    pub fn close(self) -> io::Result<()> {
-        drop(self);
-        Ok(())
+    pub fn close(self) {
+        self.shared_state.shutdown()
     }
 
     /// Calculates the round trip time between this client and the server,